requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
python-multipart
pymupdf
resend
requests
PyJWT
bcrypt
chromadb

